Ins bug fixes, code cleanup
